reimplement permission check with scap
