+++ apps/desktop/src-tauri/Cargo.toml
@@ -64,7 +64,9 @@ dotenvy_macro = "0.15.7"
 byteorder = "1.4.3"
 bytemuck = "1.14.3"
 regex = "1"
-scap = { git = "https://github.com/CapSoftware/scap", rev = "1b7c70df754ee9c2c4e040266f45d16354367bd7" }
+# scap = { git = "https://github.com/CapSoftware/scap", rev = "1b7c70df754ee9c2c4e040266f45d16354367bd7" }
+capture = { path = "../../../crates/capture" }
+scap = { path = "../../../../scap" }
 image = { version = "0.24.9", features = ["jpeg", "png", "webp"] }
 sentry = "0.32.2"
 fix-path-env = { git = "https://github.com/tauri-apps/fix-path-env-rs" }
