+++ apps/desktop/src-tauri/src/app/mod.rs
@@ -7,6 +7,8 @@ use std::{
 use tauri::{Assets, Context};
 use tracing_appender::rolling::RollingFileAppender;
 
+#[macro_use]
+pub mod commands;
 pub mod config;
 
 pub fn panic_hook(info: &PanicInfo) {
